Career_Omni/requirements.txt
streamlit>=1.37.1
unstructured>=0.15.5
pdfminer.six
pdf2image>=1.17.0
python-poppler
pillow
PyPDF2>=3.0.1
google-generativeai>=0.7.2
python-dotenv>=1.0.0
langchain>=0.2.14
langchain-community>=0.2.12
langchain-core>=0.2.33
langchain-google-genai>=0.0.11
langchain-ollama>=0.1.1
pypdf>=4.3.1
markdown-pdf>=1.3
